+++ src/plugins/boost/index.js
@@ -6,7 +6,7 @@ import routes from './routes';
 
 export default class BoostPlugin extends Plugin {
   constructor() {
-    super('triniti', 'boost', '0.2.14');
+    super('triniti', 'boost', '0.2.15');
   }
 
   configure(app, bottle) {