+++ src/plugins/ncr/index.js
@@ -16,7 +16,7 @@ import { serviceIds } from './constants';
 
 export default class NcrPlugin extends Plugin {
   constructor() {
-    super('triniti', 'ncr', '0.2.14');
+    super('triniti', 'ncr', '0.2.15');
   }
 
   configure(app, bottle) {